+++ NEWS.md
@@ -7,7 +7,7 @@ v0.2.1 (2016-??-??)
 ### New Features ###
 
 * New mathematical operations supported: `log2`, `abs2`, `fma`, `muladd`,
-  `besseli`, `besselix`, `besselk`, `besselkx`.
+  `besseli`, `besselix`, `besselk`, `besselkx`, `widen`, `big`.
 
 v0.2.0 (2016-07-13)
 -------------------